First stab at renaming this project from `ied` to `nom`
Following the discussion in alexanderGugel/ied#29 this is a first
effort to rename the project to `nom`. This would affect other
things than just this repository, such as the Travis CI set up
and surely other things as well. These changes assume those other
tools and locations have or will be changed as well.